add pan responder to bar chart
